adapt tests to overwrite functionality

+++ l2l/utils/experiment.py
@@ -90,7 +90,8 @@ class Experiment(object):
         if os.path.isdir(self.paths.output_dir_path):
             if overwrite:
                 ready_path = 'simulation/ready_files'
-                shutil.rmtree(os.path.join(self.paths.output_dir_path, ready_path))
+                if os.path.isdir(os.path.join(self.paths.output_dir_path, ready_path)):
+                    shutil.rmtree(os.path.join(self.paths.output_dir_path, ready_path))
             else: 
                 raise Exception("There are already exsiting outputfiles in this directory. Please change the path specification.")
